스냅샷 테스트 라이브러리 교체
