+++ .gitignore
@@ -136,3 +136,7 @@ iOSInjectionProject/
 # End of https://www.toptal.com/developers/gitignore/api/macos,swift,git
 Secrets.xcconfig
 GoogleService-Info.plist
+
+pinsTests/ReferenceImages_64/pinsTests.SnapshotTestTests/
+
+pinsTests/FailureDiffs/pinsTests.SnapshotTestTests/
